+++ packages/styles/src/packages/toggle/index.ts
@@ -1,0 +1,5 @@
+import * as styles from "./toggle.style.css";
+
+export const toggle = {
+  style: styles,
+};